refactor rename useInstallables -> useMods
