+++ src/mods/useModViewItems.test.tsx
@@ -26,7 +26,7 @@ import extensionsSlice from "@/store/extensionsSlice";
 import MockAdapter from "axios-mock-adapter";
 import axios from "axios";
 import { type UnavailableMod } from "@/mods/modTypes";
-import { selectUnavailableRecipe } from "@/mods/useInstallables";
+import { selectUnavailableRecipe } from "@/mods/useMods";
 import { renderHook } from "@/extensionConsole/testHelpers";
 import {
   extensionFactory,
