+++ src/mods/useMods.test.ts
@@ -16,7 +16,7 @@
  */
 
 import { renderHook } from "@/extensionConsole/testHelpers";
-import useInstallables from "@/mods/useInstallables";
+import useMods from "@/mods/useMods";
 import extensionsSlice from "@/store/extensionsSlice";
 import { validateTimestamp } from "@/types/helpers";
 import { useAllRecipes } from "@/recipes/recipesHooks";
@@ -39,7 +39,7 @@ jest.mock("@/recipes/recipesHooks", () => ({
 
 const useAllRecipesMock = jest.mocked(useAllRecipes);
 
-describe("useInstallables", () => {
+describe("useMods", () => {
   beforeEach(() => {
     appApiMock.reset();
     appApiMock.onGet("/api/extensions/").reply(200, []);
@@ -51,7 +51,7 @@ describe("useInstallables", () => {
   });
 
   it("handles empty state", async () => {
-    const wrapper = renderHook(() => useInstallables());
+    const wrapper = renderHook(() => useMods());
 
     await wrapper.waitForEffect();
 
@@ -62,7 +62,7 @@ describe("useInstallables", () => {
   });
 
   it("handles unavailable", async () => {
-    const wrapper = renderHook(() => useInstallables(), {
+    const wrapper = renderHook(() => useMods(), {
       setupRedux(dispatch) {
         dispatch(
           // eslint-disable-next-line new-cap -- unsave
@@ -94,7 +94,7 @@ describe("useInstallables", () => {
   it("multiple unavailable are single installable", async () => {
     const metadata = recipeMetadataFactory();
 
-    const wrapper = renderHook(() => useInstallables(), {
+    const wrapper = renderHook(() => useMods(), {
       setupRedux(dispatch) {
         dispatch(
           // eslint-disable-next-line new-cap -- unsave
@@ -133,7 +133,7 @@ describe("useInstallables", () => {
       error: undefined,
     } as UseCachedQueryResult<ModDefinition[]>);
 
-    const wrapper = renderHook(() => useInstallables(), {
+    const wrapper = renderHook(() => useMods(), {
       setupRedux(dispatch) {
         dispatch(
           // eslint-disable-next-line new-cap -- test setup
@@ -167,7 +167,7 @@ describe("useInstallables", () => {
   it("handles inactive cloud extension", async () => {
     appApiMock.onGet("/api/extensions/").reply(200, [cloudExtensionFactory()]);
 
-    const wrapper = renderHook(() => useInstallables());
+    const wrapper = renderHook(() => useMods());
 
     await wrapper.waitForEffect();
 
@@ -188,7 +188,7 @@ describe("useInstallables", () => {
     const cloudExtension = cloudExtensionFactory();
     appApiMock.onGet("/api/extensions/").reply(200, [cloudExtension]);
 
-    const wrapper = renderHook(() => useInstallables(), {
+    const wrapper = renderHook(() => useMods(), {
       setupRedux(dispatch) {
         dispatch(
           // eslint-disable-next-line new-cap -- test setup
